feat: Add PDF export and more info on info page, fix size listview on read page
Fixes the issues with the list view on the read page being too small and adds the ability to export the current chapter as a PDF.

Adds a new button to the read page that will generate a PDF of the current chapter.
Removes the fixed height of the list view on the read page, allowing it to grow as needed.
Adds more information to the info page, including the number of chapters and pages in the manga.
Fixes a bug where the read page would not scroll past the first page if the chapter was only one page long.
Updates the pubspec.lock file with the latest dependencies.