refactor: migrate api to use URL from SharedPreferences
This commit refactors the API to use URLs stored in SharedPreferences.

The previous approach involved using secure dotenv to load URLs from a .env file. However, this approach was problematic because the .env file was not always being loaded correctly.

To address this issue, this commit changes the API to use URLs stored in SharedPreferences. This ensures that the URLs are always available and that the API can be used without any problems.